Changing PA K-Anon threshold to 0.
Summary:
# Context

In order to align the K-Anon threshold between MPC-AEM and AEM V2, we (MPC) have changed the K-Anon threshold strategy to five clicks per day for an Ad Id as opposed to 100 conversions per Ad Id that we used previously. Thus as per this new policy, we have modified the data ingestion pipeline from Meta side to only include those Ad Object Ids in daily uploads that have at least five clicks on the given day.

In this diff, as we are handling the K-Anon check on the data ingestion pipeline on clicks and thus I need to modify the K-Anon threshold for PA to be 0.

# Change

1. Added a separate constant for PA - DEFAULT_K_ANONYMITY_THRESHOLD_PA and renamed DEFAULT_K_ANONYMITY_THRESHOLD to DEFAULT_K_ANONYMITY_THRESHOLD_PL.

2. Modified the test to check for the above change, in order to avoid duplication used parametrized.expand to pass multiple input to the unit test.

3. For point 2, added parametrized to our targets file. I see being used in several projects in fbsource and don't think there is any risk adding this additional third party dep for test target.

Differential Revision: D35237421

fbshipit-source-id: a540bc3b3b1ea8124ead83f3dff373758c05ceaa