pass binary-version tag to PC pre validation CLI
Summary:
Pass the binary_version tag to the PC pre validation CLI.

For ease of implementation, here we assume all other binaries use by the PC run will be on the same version tag as the CLI itself. This assumption does not hold for some specific use cases, where individual binary versions need to be overridden (e.g., experimentation platform).

Taking the OneDockerBinaryConfig below as an example, CLI will not honor the `ajinkyaghonge` tag but only validate the accessibility of `rc` binaries.
```
    OneDockerBinaryConfig:
      default:
        constructor:
          tmp_directory: /tmp
          binary_version: rc
          repository_path: "https://one-docker-repository-prod.s3.us-west-2.amazonaws.com/"
      private_attribution/pcf2_aggregation:
        constructor:
          tmp_directory: /tmp
          binary_version: ajinkyaghonge
          repository_path: "https://one-docker-repository-test.s3.us-west-2.amazonaws.com/"
      private_attribution/pcf2_attribution:
        constructor:
          tmp_directory: /tmp
          binary_version: ajinkyaghonge
          repository_path: "https://one-docker-repository-test.s3.us-west-2.amazonaws.com/"
```

Differential Revision: D35317611

fbshipit-source-id: fd2cbc70013af6f1ff9ce579313f61a75366760d