pass PID protocol to PID stages and combiner stage
Summary:
In this diff, we are passing PID protocol that is determined based on number of shards to each of stages.
## Determining protocol
The protocol should be determined based on the following condition.
```
if num of shards == 1:
  protocol = MULTIKEY_PID
else:
  protocol = DEFAULT_PID_PROTOCOL
```
We are determining protocol in the two places.
- `PrivateComputationServiceHandler` in measurement/private_measurement/private_computation_service/handler.py
- `create_instance()` in fbpcs/private_computation_cli/private_computation_service_wrapper.py

Q: Initially, we want to only enable this logic for Experimentation Platform. Can jrodal98 or ajinkya-ghonge give me guidance on where I should modify to do so?

## Passing protocol
Though we initially planned to add a new stage, we decided to just pass PID protocol to existing stage that are using PIDService. In order to build new stage, we need to understand PID service to replicate its functionality. We rather pass PID protocol to PIDService considering its engineering cost.

We added a new member `pid_protocol` in
- `PrivateComputationInstance`
- `PIDStage`

## PID PREPARE
Based on the protocol in `PIDPrepareStage`, PID PREPARE would switch `max_column_count` in the following manner.
```
if protocol == MULTIKEY_PID:
  max_column_count = DEFAULT_MULTIKEY_PROTOCOL_MAX_COLUMN_COUNT
else:
  max_column_count = 1
```

## PL/PA Combiner
Almost identical to PID PREPARE. In `start_combiner_service()` called from `IdSpineCombinerStageService` and `PrepareDataStageService`, it switch `max_id_column_count` in the same manner.

## PID RUN
We added PID_MULTI_KEY_SERVER and PID_MULTI_KEY_CLIENT S3 binary path in `onedocker_binary_names.py`.
In PIDProtocolRunStage, we switch binary to pass based on the protocol stored in the class.

Q: We also changed PIDStageMapper to pass something like `onedocker_binary_config_map[OneDockerBinaryNames.PID_MULTI_KEY_SERVER.value]`. I'm not sure the use of it. Can jrodal98 give me guidance?

Differential Revision: D35276714

fbshipit-source-id: d7b948b7b9a69819dc34065bf5b778ff2b99bbbb