Add deployment scripts to the GCP PCE Deployment dockerfile (#794)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/794

# Background
To deploy PCEs from the PCE service UI and CLI, we need to be able to communicate to a service within the cloud (GCP in this case) that can run the terraform scripts required to deploy a new PCE.
# This commit
This commit updates the docker image that is used to deploy PCEs to include the deployment script that is required to creat a new PCE. It also updates the docker build command to pin the built image to the linux/amd64 architecture. This became and issue in my testing as the images I built on my local machine (Mac with ARM64) wouldn't run on the nodes in the GCP clusters.

Differential Revision: D35499707

fbshipit-source-id: d8ff2279f607533524d719efe97d4bcada506d02