Update instance_status when there are possible exceptions
Summary:
**Context:**

PID Service was unable to recover from its failed state when there are uncaught failures such as ECS failures( capacity unavailable) or any other aws related exceptions. This behavior has been seen in during a canary run with Uber
post: https://fburl.com/w6yymbqu

**Reason:**

When an individual stage of PID (Ex: Shard, Prepare .. ) fails, the stage class updates the status of PIDInstance to PIDInstanceStatus.FAILED. However this happens only for known errors such as ValueError etc. If an unknown exception such as aws exceptions arise, the stage bails out without updating the PIDInstanceStatus because of which further reruns assume that the stage is still in progress and does not consider the reruns as eligible.

**Fix:**

Add a try block to catch all possible exceptions that happen during a stage run in the dispatcher and update the PIDInstanceStatus accordingly during these failures

Added test_pid_run_stage_with_exception to depict the failure scenario and retried the stage for a successful run

Reviewed By: yuyashiraki

Differential Revision: D35595625

fbshipit-source-id: dcbf8f424c32a190b2c23f74f1af3fa5c041084d