Determine PID protocol based on num shards + config
Summary:
# Description
In this diff,
- we modified `_build_pid_service()` to pass PID protocol based on number of shards and `multikey_enabled` flag in the config

# background on this change
When config or multikey_enabled flag is missing, we default back to `multikey_enabled=false` for now.
After we finish the change on WWW to force advertisers who have not onboarded to multi-key to use 2 or more shards, we will change the default to `multikey_enabled=true`. So only when it is explicitly specified, we will use multi-key for single-shard runs.

Reviewed By: rajprateek, joe1234wu

Differential Revision: D35603603

fbshipit-source-id: e8af887e68d400a09748ad4e7450407fea467354