Privately share reach in InputProcessor
Summary:
We add a method to privately share the reach in the InputProcessor. A reach occurs when the number of impressions is nonzero, and is only computed for the test population.

For details about the correctness tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit#gid=0

Reviewed By: adshastri

Differential Revision: D35789048

fbshipit-source-id: dab0e4606e46ce0546fce602bde35934f4ec8fe6