Privately share reach in InputProcessor (#831)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/831

We add a method to privately share the reach in the InputProcessor. A reach occurs when the number of impressions is nonzero, and is only computed for the test population.

For details about the correctness tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit#gid=0

Reviewed By: adshastri

Differential Revision: D35789048

fbshipit-source-id: 7cd06668a5f1b7621a86bde0b80e4c0cc5abf48d