Add main to run plain text lift calculator
Summary:
This stack of diff aims to build a binary of the plain text lift calculator to be run in a TEE environment.

This diff introduces the main file that takes in both publisher and partner inputs and uses plain-text lift calculator for computation, and writes the output to file.

Reviewed By: chualynn

Differential Revision: D35976158

fbshipit-source-id: 5f8740775f6b928c4766215483b04a4107f55768