Calculate values and reachedValues in Attributor
Summary:
We calculate the values and reachedValues in the Attributor. The value is calculated as the purchase value if there is a valid event, otherwise we set the value to be zero. Similarly, we calculate the reached value as the purchase value if there is a reach, otherwise we set it to be zero. Note that the reached value is computed only for the test population.

For more details about the correctness tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit?usp=sharing

Reviewed By: RuiyuZhu

Differential Revision: D35795503

fbshipit-source-id: d107819727f0c425c49a9e635c065f732cd66487