Privately share cohorts in InputProcessor
Summary:
We add a method to privately share the cohort group ids for both the test/control population in the InputProcessor, where the group ids are converted to index shares used for ORAM aggregation.

To compute the aggregation for both the test/control population and cohorts, we create new group ids that encode whether each row corresponds to a test/control and the cohort group id. For the test population, we use the original group id, and for the control population we add the number of cohorts to the original group id.

Note also that for the test/control population, we have to differentiate between non-opportunities, test and control. In the PL logic, we can ignore all non-opportunities, hence instead of treating them as a separate group, we treat them as part of the test group, and parse them as invalid inputs in the InputProcessor. Thus it suffices here to only share the control population, and the remaining rows are treated as the test population.

For more details about these optimizations, refer to https://docs.google.com/document/d/1frSxg9-74rtTemqmxewVA0QZAbUplsLAriBHZD-KEms/edit?usp=sharing

For more details about the correctness tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit?usp=sharing

Reviewed By: RuiyuZhu

Differential Revision: D35907556

fbshipit-source-id: c96f04abcc58a6748450200915ae0b2e097dea0c