Calculate reached conversions in Attributor (#842)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/842

We calculate the reached conversions in the Attributor. A reached conversion occurs when there is a reach (number of impressions > 0) and a valid event, and this is only computed for the test population. We pre-compute the test reach in plaintext (D35789048 (https://github.com/facebookresearch/fbpcs/commit/6b120441fb51ef4fb5c4650804c666b08334bd75)) so here we just have to take the AND of the reach with the events.

For more details about the correctness test cases, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit?usp=sharing

Reviewed By: RuiyuZhu

Differential Revision: D35795393

fbshipit-source-id: 70eb8308afb8be21d5470288fc8cd3816082643f