Calculate values and reachedValues in Attributor (#844)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/844

We calculate the values and reachedValues in the Attributor. The value is calculated as the purchase value if there is a valid event, otherwise we set the value to be zero. Similarly, we calculate the reached value as the purchase value if there is a reach, otherwise we set it to be zero. Note that the reached value is computed only for the test population.

For more details about the correctness tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit?usp=sharing

Reviewed By: RuiyuZhu

Differential Revision: D35795503

fbshipit-source-id: ae6a546c8fe70c4ef09b5eee0708559f1eec554d