Calculate value squared in Attributor (#846)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/846

We calculate the value squared in the Attributor. This is the squared sum of the values for valid events in each row. These squared sums are already pre-computed in plaintext in the InputData class, so the main logic here is to check for the first valid event and to use the squared sum corresponding to its position in the row.

For more details about the correctness tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit?usp=sharing

Reviewed By: RuiyuZhu

Differential Revision: D35795627

fbshipit-source-id: c1a3c603f4efe87ebc7849baafc1028537e55d78