Privately share test cohorts in InputProcessor (#845)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/845

We add a method to privately share index shares of the cohort group ids for the test population only. This is used for doing ORAM aggregation for metrics that are computed only for the test population, i.e. the reached conversions and the reached values.

Similarly to the new cohort group ids for the test/control population in the previous diff, here we construct new group ids, where for the test population we use the original group ids, and for the control we set the group ids to be the number of partner cohorts.

For more details about these optimizations, refer to https://docs.google.com/document/d/1frSxg9-74rtTemqmxewVA0QZAbUplsLAriBHZD-KEms/edit?usp=sharing

For more details about the tests, refer to https://docs.google.com/spreadsheets/d/1xsIl33YDU_y7lSVD5gsiGV-ubl2x9TDYiTVRyR4dS74/edit?usp=sharing

Reviewed By: RuiyuZhu

Differential Revision: D35907790

fbshipit-source-id: f9717ee3b315ab864653b0dcceae4dc3a6112b49