Use thread pool to upload files to S3 (#819)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/819

Right now files are uploaded serially, which adds a lot of walltime when dealing with a lot of shards. Plus, as we get ready to deal with scale, this is not a good solution.

In order to propagate errors, I just followed this simple idea - https://stackoverflow.com/questions/233127/how-can-i-propagate-exceptions-between-threads. the only difference is here i use a vector.

Ideally, in the future, we would improve this in one or more ways
1) Use multipart upload instead of putting the entire object. This would allow us to upload the file in chunks and in parallel
2) Don't even write to disk, directly write to S3 using the new buffered writer functionality (D34944603). It is TBD whether this will be better than option 1
3) Switch to using async uploads instead of blocking uploads. This is probably the best option, since the SDK would handle the parallelization. But definitely worth investigating if we can achieve the best performance with this.

Reviewed By: jrodal98

Differential Revision: D35651355

fbshipit-source-id: ce6727ea1bf854ddd6baa8f1702ba598825de69f