Add a deployment script for workload identity. (#852)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/852

## Background
To functionally deploy a MPC game on GCP, they need to be able to connect with GCS to get the ad data. This requires us to be able to authenticate requests from withing the OneDocker contianers that are running the MPC binaries.

## This commit
This commit adds a workload identity deployment script to the PCE-Deployment docker image. This script can be used to add workload identity authentication to a PCE cluster.

Reviewed By: zhuang-93

Differential Revision: D35846330

fbshipit-source-id: 51b32a19dec4bcc998ca40885f0fdde066946117