Annotate stage flows with binary name
Summary:
Each mpc binary which uses the cost library has a hard coded path in s3 which it uses to write the json file. See https://fburl.com/code/bzkqk6kv. In order for the PCS coordination to know which folder to read from, it has to know which binary is run for each stage. Currently there is no easy way to get this from the StageFlow that is stored on the instance. Since the post processing handlers are run at the end we will not be able to get the binary type from the pc_instance.

This diff adds two parameters which indicate the binary uses the cost logging library as well as which GameName it is.

Differential Revision: D35908046

fbshipit-source-id: d2a30b0dd019d4301806b44a2099af6c438d75d5