Remove timestamp from s3 path (#855)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/855

# Background

PC Applications currently use the cost estimation library to write files to AWS S3 with summary of network usage and memory from within the application. These are useful outside of the insights API since they reliably tell us exactly how many resources the application is using. This stack of diffs aims to read these cost files logged to S3 and write them to a hive table in FB infra. This will allow us to add this data to dashboards and monitor trends over time.

# This diff

In order to read the cost file from PCS code, we need to remove the extra timestamp. The start time of the binary is non-deterministic so we should just use the run name in the file path.

Reviewed By: ashiquemfb

Differential Revision: D35908383

fbshipit-source-id: fc68508e321b21054d2c00ddf3a1844c004a37d8