Add post processing handler to transfer cost estimation files from s3 to hive (#865)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/865

# Background

PC Applications currently use the cost estimation library to write files to AWS S3 with summary of network usage and memory from within the application. These are useful outside of the insights API since they reliably tell us exactly how many resources the application is using. This stack of diffs aims to read these cost files logged to S3 and write them to a hive table in FB infra. This will allow us to add this data to dashboards and monitor trends over time.

# This diff

In order to read the s3 files written by the cost estimation library and to write them to hive, we need to add a stage to PCS that runs after the binaries have finished. This handler will be run on FB infra and have access to the AWS accounts through the Storage Service.

In this diff we set up a new post processing handler which iterates through the stages in the stage flow and checks for each stage whether it needs to read the file from s3.

I've moved the run names which are passed into the binaries into a shared attribute on the pc_instance to keep them synced when the file is read.

Reviewed By: jrodal98

Differential Revision: D36140878

fbshipit-source-id: 25a74971885414907692fb18e948d977ab39312a