Add main file
Summary: We add a main file for running PL. For consistency with the original (EMP-based) PL, we keep the command line options and the processing of the input/output filenames to be the same. The main change is in how we create CalculatorApps accounting for concurrency and multiple files per thread, which is done in the same way as in the PCF2 attribution/aggregation apps.

Reviewed By: RuiyuZhu

Differential Revision: D36253018

fbshipit-source-id: b7a8fdb0e03f769a08abf298e0df1ac0abc7bdca