Combine Workload Identity and PCE deployment (#923)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/923

## Background
Previously we were creating a single workload identity, service account, and namespace for running onedocker containers for each PCE. This posed a potential security issue as it would give all the same permissions to each environment.

## This Commit
This commit updates the PCE deployment terraform configurations to deploy a service account, namespace and workload identity per PCE. This will allow us to give more fine grained permissions to a given environment and limit the bucket access that it has.

Reviewed By: zhuang-93

Differential Revision: D36520578

fbshipit-source-id: 5ef05e53a5d0d3b07ad9a38a2578d181ac42d1c3