Fix error message at end of ExPlat runner (#964)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/964

Currently the ExPlat runner tries to print out the arguments passed to it at the end of the run for debugging purposes. This throws an exception if you use the `--stage_flow` argument since the schema will convert the string into a `PrivateComputationBaseStageFlow`. We are using `json.dumps` to encode the object into a string, but it doesn't know how to serialize `PrivateComputationBaseStageFlow`

The fix is to add a custom JSON encoder that knows how to serialize the `PrivateComputationBaseStageFlow`

Reviewed By: adshastri

Differential Revision: D36780911

fbshipit-source-id: f9b13606c7e172eed7fb070f5f2ac7305a414c8c