Supply pid configs to MR stage service (#974)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/974

As part of diff series to enable experimentation with multi node multi key protocol, We need to pass the pid configs to private computation service through the handler.

Note that the publisher side configs are not passed in prod as they are picked from PCE config instead. The config is only passed for experimentation when we spin local thrift server while running mpc-aem-experimentation.

Reviewed By: wenqingren

Differential Revision: D36798870

fbshipit-source-id: fe5e6ca8bcf5a4b6ebc4eb5a933aaf66e033bfae