Use regular functions in AttributionRule. (#1004)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1004

Previously AttributionRule.h defined multiple std::functions to which we bind lambdas from the initialization of AttributionGame class.
The series of unnamed lamdbas in a collection (of attribution rules) initialization make the code difficult to read.

After discussling with Puhan, we decided to change it to a proper interface that any Attribution Rule can implement and created named-functions for the same.

This removes a std::function initialization in favor of a virtual function interface.

This is purely a refactoring change which should not have any logical changes.

Differential Revision: D36929271

fbshipit-source-id: 84879c96435d24ad0de8195d9891973767cf0bdb