freeze individual fields for instance_base and subclass (#1030)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1030

I am trying to freeze individual field of InstanceBase and its subclass.

We store the mutability info of each field in the matched Field object as a dictionary.

If a filed is mutable, we can always change it.
If a filed is immutable, we cannot change it after it has been initialized. An immutable field can be None, empty. It can be initialized after creating the object.

For now, if a immutable field is a reference type, we can still modify its content.

Reviewed By: gorel

Differential Revision: D36947666

fbshipit-source-id: 128779cac681fa5fb4c484327e16d56ef365158b