freeze individual fields for instance_base and subclass
Summary:
I am trying to freeze individual field of InstanceBase and its subclass.

We store the mutability info of each field in the matched Field object as a dictionary.

If a filed is mutable, we can always change it.
If a filed is immutable, we cannot change it after it has been initialized. An immutable field can be None, empty. It can be initialized after creating the object.

For now, if a immutable field is a reference type, we can still modify its content.

Differential Revision: D36947666

fbshipit-source-id: 24e855048fd32bf7efbd13850a2f830fa6d0d9e4