Develop start_pid_prepare_service() to call in PIDPrepareStageService
Summary:
In this diff, I implemented function start_pid_prepare_service() as a step to deprecate the old system PIDPrepareStage. This function takes care of the logic to spin up containers in AWS.
Function Details:
set up the arguments for the containers
uses PIDPrepareBinaryService to generate commands from the arguments
calls start_containers in PIDPrepareBinaryService, and return those containers
And it will be called by the function run_async() in  PIDPrepareStageService.

Differential Revision: D36922117

fbshipit-source-id: 0ea9dac009ec9d389feb8f1b7e10d626a131af46