E2E Test for PIDPrepareStageService (#1049)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1049

We are migrating PIDService to StageFlow, and the second step is use PIDPrepareStageService in PID_PREPARE stage.
In order to run E2E test for PIDPrepareStageService, we use the new StageFlow called PrivateComputationPIDMigrationTestStageFlow, which has the similar stage flow as PrivateComputationStageFlow but the ID_MATCH stage uses  PIDRunProtocolStageService and PID_PREPARE stage uses PIDPrepareStageService.

Reviewed By: rajprateek, joe1234wu

Differential Revision: D37101178

fbshipit-source-id: 9fe14a366956d7331255a12eead57d6ed4653816