Develop start_pid_prepare_service() to call in PIDPrepareStageService
Summary:
In this diff, I implemented function start_pid_prepare_service() as a step to deprecate the old system PIDPrepareStage. This function takes care of the logic to spin up containers in AWS.
Function Details:
set up the arguments for the containers
uses PIDPrepareBinaryService to generate commands from the arguments
calls start_containers in PIDPrepareBinaryService, and return those containers
And it will be called by the function run_async() in  PIDPrepareStageService.

Differential Revision: https://www.internalfb.com/diff/D36922117?entry_point=27

fbshipit-source-id: 524c93a9906798b302140593a9becda1a10bf157