Develop start_pid_shard_service() to call in PIDShardStageService (#1090)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1090

In this diff, I implemented function start_pid_shard_service() as a step to deprecate the old system PIDShardStage. This function takes care of the logic to spin up containers in AWS.
Function Details:
set up the arguments for the containers
uses PIDShardBinaryService to generate commands from the arguments
calls start_containers in ShardingService, and return those containers
And it will be called by the function run_async() in  PIDShardStageService.

Reviewed By: yuyashiraki

Differential Revision: D37133007

fbshipit-source-id: dc113746f28f46572aab7561004e3e27c6548f8d