Develop start_pid_shard_service() to call in PIDShardStageService
Summary:
In this diff, I implemented function start_pid_shard_service() as a step to deprecate the old system PIDShardStage. This function takes care of the logic to spin up containers in AWS.
Function Details:
set up the arguments for the containers
uses PIDShardBinaryService to generate commands from the arguments
calls start_containers in ShardingService, and return those containers
And it will be called by the function run_async() in  PIDShardStageService.

Differential Revision: https://www.internalfb.com/diff/D37133007?entry_point=27

fbshipit-source-id: 48a398514498080f9036682f6e44796c05f005fd