E2E test for PIDShardStageService
Summary:
We are migrating PIDService to StageFlow, and the last step is use PIDShardStageService in PID_SHARD stage.
In order to run E2E test for PIDShardStageService, we use the new StageFlow called PrivateComputationPIDMigrationTestStageFlow, which has the similar stage flow as PrivateComputationStageFlow but the PID_SHARD uses PIDShardStageService, the PID_PREPARE stage uses PIDPrepareStageService and the ID_MATCH stage uses  PIDRunProtocolStageService.

Reviewed By: jrodal98

Differential Revision: D37164097

fbshipit-source-id: c606d68a17efa4d14ef571a9e7bf600c42cf47af