E2E Test for PIDPrepareStageService (#1049)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1049

We are migrating PIDService to StageFlow, and the second step is use PIDPrepareStageService in PID_PREPARE stage.
In order to run E2E test for PIDPrepareStageService, we use the new StageFlow called PrivateComputationPIDMigrationTestStageFlow, which has the similar stage flow as PrivateComputationStageFlow but the ID_MATCH stage uses  PIDRunProtocolStageService and PID_PREPARE stage uses PIDPrepareStageService.

Reviewed By: rajprateek, yuyashiraki, joe1234wu

Differential Revision: D37101178

fbshipit-source-id: 68ac04535f06cc4cf350146a612018476b8d4fb3