Develop start_pid_prepare_service() to call in PIDPrepareStageService (#1078)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1078

In this diff, I implemented function start_pid_prepare_service() as a step to deprecate the old system PIDPrepareStage. This function takes care of the logic to spin up containers in AWS.
Function Details:
set up the arguments for the containers
uses PIDPrepareBinaryService to generate commands from the arguments
calls start_containers in PIDPrepareBinaryService, and return those containers
And it will be called by the function run_async() in  PIDPrepareStageService.

Reviewed By: rajprateek

Differential Revision: D36922117

fbshipit-source-id: 1cd20d789666b025ada7d116edfc1f6c6c3bbc40