E2E test for PIDShardStageService (#1092)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1092

We are migrating PIDService to StageFlow, and the last step is use PIDShardStageService in PID_SHARD stage.
In order to run E2E test for PIDShardStageService, we use the new StageFlow called PrivateComputationPIDMigrationTestStageFlow, which has the similar stage flow as PrivateComputationStageFlow but the PID_SHARD uses PIDShardStageService, the PID_PREPARE stage uses PIDPrepareStageService and the ID_MATCH stage uses  PIDRunProtocolStageService.

Reviewed By: jrodal98

Differential Revision: D37164097

fbshipit-source-id: 9b1b452fdb877a1fbf880e1fc1b74e21b9e40f5c