Fix the issue that is causing the end of deploys to fail
Summary:
When the deployment helper was made pyre-strict in D36983092 (https://github.com/facebookresearch/fbpcs/commit/60282eabe12b87379397f47780ab7f72842284a9), it introduced a
regression where the dependencies could not be resolved because the files got
moved. The fix is to create the same module structure as the one under fbpcs so
that the files can be imported as before.

Differential Revision: D37353741

fbshipit-source-id: 9ff8494e8521865475f58eeedd4b86751e217297