adding stage service stop_service for cancel current stage calls (#1117)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1117

## Why
we currently only support MPCInstance and PIDInstance to stop instance during cancel current stage.
After we introduced StageStateInstance, and the goal is to move all intances to StageStateInstance, we need to support all generic type stop action for PC service to call.
In Mao's intern project, she is migrating PIDInstances to StageStateInstance, so in this diff is to support stop container logic for generic stagestate.
## What
* Adding one new `stop_service` in StageService, allow StageService to implement it's own tear down logic
* Adopt StageService stop_service call duing cancel stage
* Add UT.

Differential Revision: D37292716

fbshipit-source-id: 9f547882cd5baa61198024fabdc31a9864052121