Share number of groups for aggregation
Summary:
We share the number of groups for computing aggregation. We will aggregate over 3 types of groups, namely the test/control population, the partner cohorts and the publisher breakdowns. To do the ORAM aggregation efficiently, we will generate a combined group id that encodes all 3 groups, which will be done in subsequent diffs.

We also do two types of aggregation, one where we consider both test/control population, and another where we only consider the test population. The latter is for the reached metrics. Hence we share the number of groups for both types of aggregation as `numGroups_` and `numTestGroups_` respectively.
- `numGroups_` is the product of the number of populations (2), the number of cohorts and the number of breakdowns.
- `numTestGroups_` is the product of the number of cohorts and the number of breakdowns for the test population, and we add 1 group for the control population (disregarding the cohorts/breakdowns).

Reviewed By: gorel

Differential Revision: D37337932

fbshipit-source-id: c77e0653fc0ea8e6b478b964c52f4b5ac0378989