Share publisher breakdown ids (#1142)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1142

We share the publisher breakdown ids and the number of publisher breakdowns in the InputProcessor class. Since the breakdown id is either 0 or 1, we parse them as booleans in the InputData class, and store them using the SecBit type in the InputProcessor. They will be used for aggregating the metrics depending on the breakdown id in the subsequent diffs.

Reviewed By: gorel

Differential Revision: D37337517

fbshipit-source-id: 76c5fd7f9798700fd89de29b39d0bfe4d18cab53