Refactor methods for revealing cohort and population output (#1155)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1155

We refactor the methods in the Aggregator for revealing the cohort and population aggregation output. Previously, we combined them into a two methods for aggregating over both, depending on whether we are only aggregating over the test population or whether we are aggregating over both test and control populations. We refactor them here because we will be adding aggregation over the publisher breakdowns as well in the subsequent diffs.

Reviewed By: RuiyuZhu

Differential Revision: D37338465

fbshipit-source-id: e0e7b7fb05b413f1b7e031e86f7d9dbc2a68dbde