Construct index shares including breakdowns (#1156)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1156

We construct index shares for doing ORAM aggregation over the populations, cohorts and breakdowns. Previously, we only considered the population and cohorts, and in this diff we add logic to include breakdowns as well. To do this, we encode the three types of groups into a single group id, which we convert to boolean index shares that are input to the ORAM.

Since there are at most 2 populations (test/control) and 2 breakdowns (0/1), there are up to 4 * numPartnerCohorts_ group ids in total. We assign the first 2 * numPartnerCohorts_ group ids to the test population, and the second half to the control population. Within the test population, we assign the group ids 0 to numPartnerCohorts_ - 1 to breakdown id 0, and the group ids from numPartnerCohorts_ to 2 * numPartnerCohorts_ - 1 to breakdown id 1. We similarly assign the group ids for the control population.

For the case where we only consider the test population, we assign the all group ids corresponding to the control as a single group id, while the logic for assigning test group ids remains the same.

Reviewed By: RuiyuZhu

Differential Revision: D37367013

fbshipit-source-id: 3df38cc41083abee46514941ec8f06b152c4ca0e