Add padding size as input to PCF2 PL (#1222)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1222

We add the padding size as an input to the PCF2 PL stage, in order to specify whether to run basic or advanced lift. The command line flag for the padding size in the PL binary is called `num_conversions_per_user`. The current default is 25 for advanced lift, and to run basic lift, we can specify the padding size to be 4.

Reviewed By: gorel

Differential Revision: D36795001

fbshipit-source-id: 48a2363f6d32641ba4c1f15cd74cb51d622462f5