Centralize PID Related Variables (#1253)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1253

Why?
Currently, those variables are fragmented in different places. For example, mltikey_enabled is an attribute of PIDService which will be deprecated soon. And pid_use_row_number, and pid_config are attributes within pc_instance. What’s more, the PID protocol and max_col_cnt are not stored as attributes and we have to calculate them repeatedly in different PID stages.
So before deleting the PIDService, we centralized those variables in one place to increase code maintenance and decrease redundancy.

Detail:
In this diff, we instantiate the following variables when instantiating PrivateComputationInstance:
multikey_enabled
pid_protocol
max_column_count

Reviewed By: gorel

Differential Revision: D37699936

fbshipit-source-id: 35885d6e977cb9b366ed8ca907add4627057f0b8