Add new aggregateMetric for shard combiner
Summary:
# context
Contains the AggMetric header to support PCF2.0 computations.
This is version of the AggMetric.h is intended to be more generic to support XOR-SS, Arithmetic-SS or anything else. I have added comment to reflect this.
# Note
Splitting the diff. I'm not including this file in the targets file, will add later.

# Changes in this diff
1. Add new AggMetric header file.

Reviewed By: chualynn, ajinkya-ghonge

Differential Revision: D37618435

fbshipit-source-id: 7d4b4cb89c8ac37e5f00f6a0f82ab1e4e374bf6f