Add new aggregateMetric for shard combiner (#1269)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1269

# context
Contains the AggMetric header to support PCF2.0 computations.
This is version of the AggMetric.h is intended to be more generic to support XOR-SS, Arithmetic-SS or anything else. I have added comment to reflect this.
# Note
Splitting the diff. I'm not including this file in the targets file, will add later.

# Changes in this diff
1. Add new AggMetric header file.

Reviewed By: chualynn, ajinkya-ghonge

Differential Revision: D37618435

fbshipit-source-id: b0b9780cc7515278aed1fe1cc56f7113aae31176