Set FBPCS_IMAGE_DIGEST sha in run_fbpcs.sh, add image digest in the log for easy troubleshooting
Summary:
## Why
During onboarding triage, it's hard to tell what's the version of the image bundle in the partner side's run.
for example, in previous posts https://fb.workplace.com/groups/331044242148818/permalink/525450086041565/
https://fb.workplace.com/groups/331044242148818/permalink/521810729738834/

It's a pain point for us to check whether partners are pulling the same docker image as publisher side tiers.

We want to see the image inspect in the log to help engineers to better troubleshooting
## What
Including image digest in the partner side log which helped us to narrow down the version of bundle caused the issues

* get image digest on shell script
** https://stackoverflow.com/questions/32046334/where-can-i-find-the-sha256-code-of-a-docker-image
* passing to docker run w/ `-e` to set env variable
* in CLI, get OS env, and print out in the log

Differential Revision: D37829033

fbshipit-source-id: d7092d69065a3b78c5fe655103a5eecd80ecdb89