create update_generic_hook for status_update_ts
Summary:
# Why:
The logic is: `status_update_ts` will record the time when status is updated. This was in `update_status` function in `PraviteComputationInstance.py`, After we have officially decoupled infra and product attributes, it will be much better to apply this in `InfraConfig` since both `status` and `status_update_ts` are in `InfraConfig` now.

# What:
create a `UpdateGenericHook` for `status`. Whenever `status` is updated, update `status_update_ts`. I am not using `UpdateOtherFieldHook` (Updates some *other* field in this instance whenever this field gets updated) here, because there are more to add in `post_update_status` function.

Reviewed By: gorel

Differential Revision: D37620672

fbshipit-source-id: dcd7748428fbb78e1613c06992d1300315129634