add new status_updates in infra_config
Summary:
# What:
A `StatusUpdates` object will be used to keep track of the `status` and its update time. This `status_updates` will be updated whenever `status` is updated. I add this in `post_status_update` function, make it part of an UpdateGenericHook: `post_status_hook`.

This attribute is passed in as an empty list when creating a `PrivateComputationInstance` at the beginning of a lift/attribution run.

Since this new attribute is introduced in `InfraConfig`, we need to add a new pass-in in many test files.

Differential Revision: D37598461

fbshipit-source-id: 9a4f6ba5398c92b727aaa8e0f22fc17461edbb24