create update_generic_hook for status_update_ts
Summary:
# why:
The logic is: status_update_ts will record the time when status is updated. This was in update_status function in PraviteComputationInstance.py, After we have officially decoupled infra and product, it will be much better to apply this in InfraConfig since both status and status_update_ts are in InfraConfig now.

# What:
create a UpdateGenericHook for stauts. Whenever status is update, update status_update_ts. I am not using UpdateOtherFieldHook (Updates some *other* field in this instance whenever this field gets updated) here, because there are more to add in post_update_status function.

Differential Revision: D37620672

fbshipit-source-id: 12c739b3d2c0effa88c12a6e687980d74fc01796