add new status_updates in infra_config (#1279)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1279

# What:
A `StatusUpdate` object will be used to keep track of the `status` and its update time. This `status_updates` will be updated whenever `status` is updated. I add this in `post_status_update` function, make it part of an UpdateGenericHook: `post_status_hook`.

This attribute is passed in as an empty list when creating a `PrivateComputationInstance` at the beginning of a lift/attribution run.

Since this new attribute is introduced in `InfraConfig`, we need to add a new pass-in in many test files.

Reviewed By: gorel

Differential Revision: D37598461

fbshipit-source-id: 822ac7f05817546822530cf6830991283c7026d7