Set FBPCS_BUNDLE_ID ENV in Dockefile, passing bundle id when building image
Summary:
## Why
During onboarding triage, it's hard to tell what's the version of the image bundle in the partner side's run.
for example, in previous posts
https://fb.workplace.com/groups/331044242148818/permalink/525450086041565/
https://fb.workplace.com/groups/331044242148818/permalink/521810729738834/
It's a pain point for us to check whether partners are pulling the same docker image as publisher side tiers.
We want to see the image inspect in the log to help engineers to better troubleshooting

## What
* Add FBPCS_BUNDLE_ID ARG in docker file, make ENV FBPCS_BUNDLE_ID default 0 if no `--build-arg` pass in for backward compatible
* Add `--build-arg` in Github CI/CD, no deploement depency
* in CLI, get OS env, and print out in the log

## Next
* store bundle id into pcinstance as version

Differential Revision: D37861394

fbshipit-source-id: f5650e5d7e4d92dbe594d5b58a64bad3ee9bc4d4