Fix index share computation when there are no breakdowns
Summary: Previously, our implementation of the index share computation for ORAM aggregation erroneously assumes that breakdown ids are always present. This caused a bug when we run the computation without computing breakdowns. To fix this, we change the computation of the index shares to account for the case where there are no breakdowns.

Reviewed By: RuiyuZhu

Differential Revision: D37867128

fbshipit-source-id: ed2c7c989d9f3d5ffc6b88e8ddde8b528fc93ba1