Remove async_utils.py dependency from fbpcs (#1285)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1285

in previous BE diff - D37797458 (https://github.com/facebookresearch/fbpcs/commit/ea3637b58258091c3dc66e9d4aabf6f9abeb1d83). we removed libfb dependency, which unblocked us to remove `to_sync` annotation which was used for `data_provider`
Grab the summary from T105383010

## Why
We have test code in the file mentioned in the title which references async_utils.py. This file only existed because we didn't previously have the ability to write async tests. Fortunately, the built-in unittest library has come a long way since that file was first written, so now we want to deprecate it. Before we can do that, we need to migrate the last few files using it.

## What
Modify the test file to no longer depend on async_utils.py
Many of the crutches we used to rely on are now available in python by default. For reference, anything available in python3.8 is fair game.

*  remove`to_sync` annotation which was used for `data_provider` (been removed in D37797458 (https://github.com/facebookresearch/fbpcs/commit/ea3637b58258091c3dc66e9d4aabf6f9abeb1d83))
* remove `async_util.py` file
* use `unittest.mock.AsyncMock`

Differential Revision: D37871469

fbshipit-source-id: a5d344c8cf8b6f6fd6d2adb24ce730f95cb6c0a5