Set FBPCS_BUNDLE_ID ENV in Dockefile, passing bundle id when building image (#1282)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1282

## Why
During onboarding triage, it's hard to tell what's the version of the image bundle in the partner side's run.
for example, in previous posts
https://fb.workplace.com/groups/331044242148818/permalink/525450086041565/
https://fb.workplace.com/groups/331044242148818/permalink/521810729738834/
It's a pain point for us to check whether partners are pulling the same docker image as publisher side tiers.
We want to see the image inspect in the log to help engineers better troubleshooting

This version improved the previous diff D37829033
## What
* Add `FBPCS_BUNDLE_ID` `ARG` in docker file, make `ENV` `FBPCS_BUNDLE_ID` default `0` if no `--build-arg` pass in for backward compatible
* Add `--build-arg` in Github CI/CD, no deployment dependency
* in CLI, get OS env, and print out in the log

## Next
* store bundle id into PC instance as version - EveChenC

Reviewed By: gorel

Differential Revision: D37861394

fbshipit-source-id: 11b7b1c8f19516a1705bb70e53a3fcaa8efb399b