] set end_ts immutable (#1293)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1293

# What:
`end_ts` will be updated when the `status` is completed. This update of `end_ts` was in `update_status` function in `PrivateComputationInstance`. The functions used to check if `status` is completed were also in `PrivateComputationInstance`.
I maintain the APIs in `PrivateComputationInstance`, but move the actual code into `InfraConfig` since they should be part of infra. This can be considered as part of the decoupling work. This move leads to a reduction of the times we call `time.time()` -> so I changed 2 to 1 in `test_private_computation.py`.
I add this completed-check-update in `post_update_status` function, make it part of `post_status_hook`.
I create a `FrozenFieldHook` for `end_ts`. This field was initilized as 0. So when it is updated and not 0, it is the time to set it as immutable.

Reviewed By: gorel

Differential Revision: D37773574

fbshipit-source-id: c5f320d27b0bac48b9531cb6287d9d09b2de22b3